% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_3d_layers.R
\name{add_direction}
\alias{add_direction}
\title{Add a line of predicted effects to the plotly surface, with optional confidence intervals. User can add a line in any direction.}
\usage{
add_direction(
  p,
  model,
  direction_data,
  direction_name = "User defined line",
  linecolor = "black",
  ci = T
)
}
\arguments{
\item{p}{A plotly object}

\item{model}{A glm with exactly two x variables}

\item{direction_data}{A data frame with of x values, predicted y values and predicted confidence interval for each pair of x values. The variable names are c("rownum", actual x1 variable name, actual x2 variable name, actual y variable name, "lowerCI", "upperCI")}

\item{direction_name}{The hover text for the plotted line(s). Defaults to "User defined line".}

\item{linecolor}{The color for the plotted line. Defaults to "black".}

\item{ci}{An optional logical. Defaults to TRUE, showing the confidence intervals of the predicted effects.}
}
\value{
A plotly object
}
\description{
Primarily used by functions such as add_3d_surface or add_marginals.
If user defines "direction_data" appropriately, any line can be shown.
}
