% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_3d_layers.R
\name{add_marginals}
\alias{add_marginals}
\title{Add 3d marginal effects to a plot_ly plot}
\usage{
add_marginals(
  p,
  model,
  data = NULL,
  ci = T,
  x1_constant_val = "mean",
  x2_constant_val = "mean",
  x1_color = "darkorange",
  x2_color = "crimson",
  x1_direction_name = "Predicted marginal effect of x1",
  x2_direction_name = "Predicted marginal effect of x2",
  omit_x1 = F,
  omit_x2 = F
)
}
\arguments{
\item{p}{A plotly object}

\item{model}{A glm with exactly two x variables}

\item{data}{An optional dataframe to be used to create the regression surface. By default, this will be the data used by the inherited plotly object.}

\item{ci}{An optional logical. Defaults to TRUE, showing the confidence intervals of the predicted effects.}

\item{x1_constant_val}{A string or numeric value indicating which constant value to set for x1 when the marginal effect of x2 is plotted. Defaults to the mean value. The string can take on "mean", "median", "min", or "max". Alternately, a numeric value may be specified.}

\item{x2_constant_val}{A string or numeric value indicating which constant value to set for x2 when the marginal effect of x1 is plotted. Defaults to the mean value. The string can take on "mean", "median", "min", or "max". Alternately, a numeric value may be specified.}

\item{x1_color}{The color to be used for the line(s) depicting the marginal effect of x1. Defaults to "darkorange".}

\item{x2_color}{The color to be used for the line(s) depicting the marginal effect of x2. Defaults to "crimson".}

\item{x1_direction_name}{The hover text for the plotted line(s). Defaults to "Predicted marginal effect of x1".}

\item{x2_direction_name}{The hover text for the plotted line(s). Defaults to "Predicted marginal effect of x2".}

\item{omit_x1, omit_x2}{An optional logical. Defaults to FALSE. If set to TRUE, the marginal effect for that variable will not be included.}
}
\value{
A plotly object with the predicted marginal effects added to the plot.
}
\description{
Additional plotly layers such as add_markers() can
be added to the plotly plot, but be aware that many plotly layers inherit the data from the prior layer.
As such, a function such as add_markers may not work as intended if called after add_marginals().
}
