% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_data.R
\name{create_constant_value}
\alias{create_constant_value}
\title{Create a constant value}
\usage{
create_constant_value(constant_value, data, coefficients, x_thats_constant)
}
\arguments{
\item{constant_value}{The desired constant value for the marginal effect. It can take the value "mean", "median", "min", "max", or a numeric value}

\item{data}{The data frame being used for the regression.}

\item{coefficients}{A named character vector where the names are the c("y","x1","x2") and the values are the variable names associated with each element in the named vector.}

\item{x_thats_constant}{The variable used to create the constant value.}
}
\value{
A numeric value corresponding to the constant value desired (e.g. the mean, median, etc)
}
\description{
Create a constant value
}
