% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_data.R
\name{create_glm_adjustment}
\alias{create_glm_adjustment}
\title{Create predicted and confidence interval values from a glm model.}
\usage{
create_glm_adjustment(x_vals, model)
}
\arguments{
\item{x_vals}{A data frame or tibble with x values.}

\item{model}{A glm with exactly two x variables}
}
\value{
A data frame or tibble with x values and their corresponding predicted y values.
This is an intermediary function.
create_y_estimates generates the appropriate column names and data structure.
}
\description{
Used by create_y_estimates when the model is a glm.
This has been tested on the binomial family, but may work for all glms.
Uses the inverse link function to generate both predicted values and confidence intervals for the range of x values.
}
