% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_data.R
\name{create_marginal_x_vars}
\alias{create_marginal_x_vars}
\title{Create a data frame of x variables to plot the marginal effect}
\usage{
create_marginal_x_vars(data, model, marginal_of_x1, constant_value = "mean")
}
\arguments{
\item{data}{A data frame or a tibble (check it works with a tibble)}

\item{model}{A glm with exactly two x variables}

\item{marginal_of_x1}{A logical. If true, then the data frame represents the marginal effect of x1. If false, the marginal effect of x2.}

\item{constant_value}{A string or number indicating which constant value to use for the variable that remains constant. The string can take on "mean", "median", "min", or "max". Alternately, a numeric value may be specified.}
}
\value{
A data frame of x values that span the minimum to maximum values of the marginal effect, holding the other x variable constant
}
\description{
Create a data frame of x variables to plot the marginal effect
}
