% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_data.R
\name{create_named_coeffs}
\alias{create_named_coeffs}
\title{Create named character vector attaching coefficient names to standardized names}
\usage{
create_named_coeffs(model)
}
\arguments{
\item{model}{A glm with exactly two x variables}
}
\value{
A named character vector attaching coefficient names to standardized names (e.g. x1, x2, y)
}
\description{
Create named character vector attaching coefficient names to standardized names
}
\examples{
mymodel <- lm(length ~ isFemale_num + isMale_num, data = hair_data)
coefficients <- create_named_coeffs(model= mymodel)
}
