% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_data.R
\name{create_surface_x_vars}
\alias{create_surface_x_vars}
\title{Create matrix of x variables for regression surface}
\usage{
create_surface_x_vars(data, model, coefficient_names)
}
\arguments{
\item{data}{A data frame or a tibble (check it works with a tibble)}

\item{model}{A glm with exactly two x variables}

\item{coefficient_names}{A named character vector that attaches coefficient names to standardized names (e.g. x1, x2, y)}
}
\value{
A data frame of x values that span the minimum to maximum values of both x variables
}
\description{
Create matrix of x variables for regression surface
}
