% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_data.R
\name{create_y_estimates}
\alias{create_y_estimates}
\title{Create predicted y values from a data frame of x values}
\usage{
create_y_estimates(x_vals, model, coefficient_names)
}
\arguments{
\item{x_vals}{A data frame  or tibble with x values}

\item{model}{A glm with exactly two x variables}

\item{coefficient_names}{A named character vector that attaches coefficient names to standardized names (e.g. x1, x2, y)}
}
\value{
A data frame with x values and their corresponding predicted y values
}
\description{
Create predicted y values from a data frame of x values
}
