% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hair_data}
\alias{hair_data}
\title{Simulated hair length data}
\format{
\code{hair_data}
A data frame with 1630 rows and 9 columns:
\describe{
\item{X}{Row number}
\item{gender}{Gender identification: male, female, or nonbinary}
\item{length}{Hair length in inches}
\item{isFemale_num}{numeric binary indicator for whether observation is female}
\item{isFemale}{Labels for isFemale_num}
\item{isMale_num}{numeric binary indicator for whether observation is male}
\item{isMale}{Labels for isMale_num}
\item{isNonbinary_num}{numeric binary indicator for whether observation is nonbinary}
\item{isNonbinary}{Labels for isNonbinary_num}
}
}
\source{
R/hair_data.R
}
\usage{
hair_data
}
\description{
A simulated dataset that shows plausible hair lengths for 3 gender categories.
}
\keyword{datasets}
