% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_jitter_layers.R
\name{jitter_40pct}
\alias{jitter_40pct}
\title{Create jittered data}
\usage{
jitter_40pct(data, varname, jitter_val = NULL)
}
\arguments{
\item{data}{The data frame with the x, y, and z variables.}

\item{varname}{The name of the variable to add jitter to.}

\item{jitter_val}{The amount of jitter to add. If it is NULL (default value), then the function calculates 40\% of the range of the data.}
}
\value{
A dataframe containing one jittered column.
}
\description{
Used by add_jitter()
}
