% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress3d-package.R
\docType{package}
\name{regress3d-package}
\alias{regress3d}
\alias{regress3d-package}
\title{regress3d: Create 3D Regression Surfaces}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Plot regression surfaces and marginal effects in three dimensions. The plots are 'plotly' objects and can be customized using functions and arguments from the 'plotly' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ellaFosterMolina/regress3d}
  \item \url{https://ellafostermolina.github.io/regress3d/}
  \item Report bugs at \url{https://github.com/ellaFosterMolina/regress3d/issues}
}

}
\author{
\strong{Maintainer}: Ella Foster-Molina \email{ella.fostermolina@gmail.com} (\href{https://orcid.org/0000-0001-7455-7221}{ORCID}) [copyright holder]

}
\keyword{internal}
