% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{tibble_to_dataframe}
\alias{tibble_to_dataframe}
\title{Coerce tibble to data frame, used for testing}
\usage{
tibble_to_dataframe(tibble)
}
\arguments{
\item{tibble}{A tibble}
}
\value{
A data frame.
}
\description{
Commented out the message as unnecessary. Could get rid of this function.
}
