% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_plot_data.R
\name{create_surface_data}
\alias{create_surface_data}
\title{Create data frame used to plot a surface of predicted y values}
\usage{
create_surface_data(data, model)
}
\arguments{
\item{data}{A data frame being used to create the regression model}

\item{model}{A glm with exactly two x variables}
}
\value{
A data frame with a matrix of x values, predicted y values and predicted confidence intervals for each pair of x values
}
\description{
Create data frame used to plot a surface of predicted y values
}
\examples{
mymodel <- lm(length ~ isFemale_num + isMale_num, data = hair_data)
surface_data <- create_surface_data(data = hair_data, model = mymodel)
}
